pyproject.toml
[project]
name = "neuroweave"
version = "0.1.0"
description = "Decentralized technical knowledge extraction from Discord"
requires-python = ">=3.12"
dependencies = [
    # FastAPI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.6",
    # LangGraph Pipeline
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    # ML
    "sentence-transformers>=3.3.0",
    "scikit-learn>=1.6.0",
    # Task Queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    # LangGraph Checkpointing
    "langgraph-checkpoint-mongodb>=0.1.0",
    # Discord Bot
    "discord.py>=2.4.0",
    # Observability
    "structlog>=24.4.0",
    "langsmith>=0.2.0",
    "tenacity>=9.0.0",
    # Auth
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    # C2PA
    "c2pa-python>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["api", "bot"]

[build-system]
requires = ["hatchling>=1.28.0"]
build-backend = "hatchling.build"
